src/requirements.txt
requests==2.24.0
Django==3.1
djangorestframework==3.11.1